requirements.txt
slack-bolt
python-dotenv
psycopg2-binary
pytz
flask
redis
requests
